+++ extensions/cpanel/CHANGELOG.md
@@ -1,5 +1,10 @@
 # cPanel Changelog
 
+## [View Databases + Create Limited DNS Records] - {PR_MERGE_DATE}
+
+- View MySQL or PostgreSQL Databases
+- In Domain DNS Zone, you can create `A`, `AAAA`, `CNAME`, `TXT` records
+
 ## [Initial Version] - 2024-08-01
 
 - View Domains and their DNS Zones